Unset loaded army if it's not in savedArmies
This is characteristic of it having been deleted elsewhere
To make this characterisation true, we have to distinguish it from
`savedArmies` still being in its default state, which I've done by
making and updating a separate state variable. If `savedArmies` has not
been populated yet, we return early with no changes from
`armyHasChanges`, but we don't unset the loaded army. This is different
from `savedArmies` having been populated and being empty because all
armies have been deleted.